+++ R/bind.r
@@ -117,6 +117,12 @@ bind_rows <- function(..., .id = NULL) {
       x <- compact(x)
       names(x) <- seq_along(x)
     }
+
+    walk(x, function(.x) {
+      if(.id %in% names(.x)) {
+        glubort(NULL, '"{.id}" is a column name, it cannot be used as .id')
+      }
+    })
   }
 
   bind_rows_(x, .id)