+++ inst/include/dplyr/hybrid/scalar_result/mean_sd_var.h
@@ -185,12 +185,12 @@ SEXP meansdvar_dispatch(const SlicedTibble& data, const Expression<SlicedTibble>
   switch (expression.size()) {
   case 1:
     // fun( <column> )
-    if (expression.is_unnamed(0) && expression.is_column(0, x)) {
+    if (expression.is_unnamed(0) && expression.is_column(0, x) && x.is_trivial()) {
       return internal::SimpleDispatch<SlicedTibble, Impl, Operation>(data, x, na_rm, op).get();
     }
   case 2:
     // fun( <column>, na.rm = <bool> )
-    if (expression.is_unnamed(0) && expression.is_column(0, x) &&
+    if (expression.is_unnamed(0) && expression.is_column(0, x) && x.is_trivial() &&
         expression.is_named(1, symbols::narm) && expression.is_scalar_logical(1, na_rm)
        ) {
       return internal::SimpleDispatch<SlicedTibble, Impl, Operation>(data, x, na_rm, op).get();