+++ inst/include/dplyr/hybrid/scalar_result/sum.h
@@ -119,13 +119,13 @@ SEXP sum_dispatch(const SlicedTibble& data, const Expression<SlicedTibble>& expr
   switch (expression.size()) {
   case 1:
     // sum( <column> )
-    if (expression.is_unnamed(0) && expression.is_column(0, x)) {
+    if (expression.is_unnamed(0) && expression.is_column(0, x) && x.is_trivial()) {
       return sum_(data, x, /* na.rm = */ false, op);
     }
     break;
   case 2:
     bool test;
-    if (expression.is_unnamed(0) && expression.is_column(0, x) &&
+    if (expression.is_unnamed(0) && expression.is_column(0, x) && x.is_trivial() &&
         expression.is_named(1, symbols::narm) && expression.is_scalar_logical(1, test)
        ) {
       return sum_(data, x, test, op);