+++ inst/include/dplyr/hybrid/vector_result/ntile.h
@@ -127,7 +127,7 @@ SEXP ntile_dispatch(const SlicedTibble& data, const Expression<SlicedTibble>& ex
   case 2:
     // ntile( <column>, n = <int> )
     Column x;
-    if (expression.is_unnamed(0) && expression.is_column(0, x) && expression.is_named(1, symbols::n) && expression.is_scalar_int(1, n)) {
+    if (expression.is_unnamed(0) && expression.is_column(0, x) && x.is_trivial() && expression.is_named(1, symbols::n) && expression.is_scalar_int(1, n)) {
       return ntile_2(data, x, n, op);
     }
   default: