+++ tests/testthat/test-group_map.R
@@ -97,7 +97,7 @@ test_that("group_modify() works on ungrouped data frames (#4067)", {
   )
 })
 
-test_that("group_map() uses ptype on empty splits", {
+test_that("group_map() uses ptype on empty splits (#4421)", {
   res <- mtcars %>%
     group_by(cyl) %>%
     filter(hp > 1000) %>%
@@ -107,4 +107,12 @@ test_that("group_map() uses ptype on empty splits", {
   expect_equal(names(ptype), setdiff(names(mtcars), "cyl"))
   expect_equal(nrow(ptype), 0L)
   expect_is(ptype, "data.frame")
+})
+
+test_that("group_modify() uses ptype on empty splits (#4421)", {
+  res <- mtcars %>%
+    group_by(cyl) %>%
+    filter(hp > 1000) %>%
+    group_modify(~.x)
+  expect_equal(res, group_by(mtcars[integer(0L), ], cyl))
 })