+++ R/tbl-df.r
@@ -17,7 +17,9 @@ as.tbl.data.frame <- function(x, ...) {
 }
 
 #' @export
-tbl_vars.data.frame <- function(x) names(x)
+tbl_vars.data.frame <- function(x) {
+  names(x)
+}
 
 #' @export
 same_src.data.frame <- function(x, y) {