+++ NEWS.md
@@ -1,5 +1,7 @@
 # dplyr 0.8.1.9000
 
+* `group_by()` puts NA groups last in character vectors (#4227).
+
 * `first()`, `last()` and `nth()` hybrid version handles factors (#4295).
 
 * Using quosures in colwise verbs is deprecated (#4330).