+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `tbl_vars()` keeps information on grouping columns (#4106). 
 
+* `group_by()` puts NA groups last in character vectors (#4227).
+
 * `first()`, `last()` and `nth()` hybrid version handles factors (#4295).
 
 * Using quosures in colwise verbs is deprecated (#4330).