+++ NEWS.md
@@ -2,6 +2,12 @@
 
 * `tbl_vars()` keeps information on grouping columns (#4106). 
 
+* `group_split()` always sets the `ptype` attribute. 
+
+* `group_map()` and `group_modify()` work in the 0 group edge case (#4421)
+
+* `view()` is reexported from tibble (#4423). 
+
 * `top_n()` quotes its `n` argument. 
 
 * `top_frac(data, proportion)` is a shorthand for `top_n(data, proportion * n())` (#4017).