+++ NEWS.md
@@ -1,3 +1,5 @@
+# dplyr (development version)
+
 # dplyr 0.8.2 (2019-06-28)
 
 ## New functions